created Collector subsystem class
