+++ src/main/java/com/team1816/season/subsystems/Elevator.java
@@ -1,0 +1,30 @@
+package com.team1816.season.subsystems;
+
+import com.team1816.lib.subsystems.Subsystem;
+
+public class Elevator extends Subsystem {
+    @Override
+    public void readFromHardware() {
+
+    }
+
+    @Override
+    public void writeToHardware() {
+
+    }
+
+    @Override
+    public void zeroSensors() {
+
+    }
+
+    @Override
+    public void stop() {
+
+    }
+
+    @Override
+    public boolean testSubsystem() {
+        return false;
+    }
+}
